Init commit
This has the basic functionality of the program - You can instantly save any chrome window and give it a name, allowing you to open its contents if you want to look at it later.

Also FYI this is a horribly coded mess at the moment, so observe caution when using this dumpster fire- I mean extension.